地址管理

+++ src/main/java/cn/nomeatcoder/dal/mapper/ProductMapper.java
@@ -3,7 +3,6 @@ package cn.nomeatcoder.dal.mapper;
 import cn.nomeatcoder.common.Mapper;
 import cn.nomeatcoder.common.domain.Product;
 import cn.nomeatcoder.common.query.ProductQuery;
-import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
@@ -15,7 +14,7 @@ import java.util.List;
  * @Date 2020-01-31 14:38:45
  */
 public interface ProductMapper extends Mapper<ProductQuery, Product> {
-	long delete(ProductQuery query);
+	int delete(ProductQuery query);
 	List<Product> selectByNameAndProductId(ProductQuery query);
 	List<Product> selectByNameAndCategoryIds(ProductQuery query);
